refactor(async): swap legacy Tornado gnrasync, drop tornado_wsgi
- gnrasync_aio.py renamed to gnrasync.py: the asyncio/aiohttp port
  is now the canonical module. Public symbols (GnrAsyncServer,
  GnrBaseAsyncServer, SharedObject, SqlSharedObject, SharedLogger,
  SharedStatus, SharedObjectsManager, DelayedCall, threadpool,
  lockedCoroutine, lockedThreadpool, AsyncWebSocketHandler) match
  the legacy import surface so all callers keep working.
- tornado_wsgi.py deleted. The Tornado-serves-WSGI mode (gnr web
  serve --tornado, GnrAsyncServer(web=True)) is gone. WSGI dev
  server is werkzeug; production WSGI is gunicorn (gnr web stack).
  The gnrasync server now only handles WebSocket, /wsproxy push,
  and the pdb debugger TCP bridge.
- serverwsgi.py: --tornado flag dropped. Replaced with --async-too,
  which spawns a separate gnrasync child process alongside the
  werkzeug dev server (subprocess.Popen with atexit cleanup). The
  old in-process Tornado web=True path is removed and the rest of
  the dev-mode flow runs unconditionally. Skipped under werkzeug
  reloader child to avoid double-spawn.
- gnrwsgisite.py: the dead 'tornado=None' kwarg on GnrWsgiSite is
  removed (it was accepted but never read inside the class).
